+++ lib/src/internal/clamp.dart
@@ -1,3 +1,3 @@
-int clamp(int x, int a, int b) => x.clamp(a, b);
+int clamp(int x, int a, int b) => x.clamp(a, b).toInt();
 
-int clamp255(int x) => x.clamp(0, 255);
+int clamp255(int x) => x.clamp(0, 255).toInt();
